+++ src/components/RecipeForm/RecipeForm.jsx
@@ -1,0 +1,300 @@
+import React from "react";
+import { X } from "lucide-react";
+
+import { useRecipeForm } from "../../hooks/UseRecipeForm";
+import "./RecipeForm.css";
+
+const RecipeForm = ({
+  categories,
+  initialRecipe = null,
+  title = "Add New Recipe",
+}) => {
+  const {
+    formData,
+    validationErrors,
+    loading,
+    error,
+    isEditMode,
+    handleInputChange,
+    handleIngredientChange,
+    handleInstructionChange,
+    addInstruction,
+    removeInstruction,
+    addIngredient,
+    removeIngredient,
+    handleEnter,
+    handleSubmit,
+    handleCancel,
+    toTitleCase,
+  } = useRecipeForm(initialRecipe);
+
+  const unitOptions = [
+    { value: "", label: "Unit", disabled: true },
+    { value: "tsp", label: "tsp" },
+    { value: "tbsp", label: "tbsp" },
+    { value: "cup", label: "cup" },
+    { value: "cups", label: "cups" },
+    { value: "ml", label: "ml" },
+    { value: "g", label: "g" },
+    { value: "kg", label: "kg" },
+    { value: "can", label: "can" },
+    { value: "cans", label: "cans" },
+    { value: "piece", label: "piece" },
+    { value: "pieces", label: "pieces" },
+    { value: "clove", label: "clove" },
+    { value: "cloves", label: "cloves" },
+  ];
+
+  return (
+    <div className="add-recipe-page">
+      <div className="add-recipe-container">
+        <header className="page-header">
+          <h1>{title}</h1>
+        </header>
+
+        <form onSubmit={handleSubmit} className="recipe-form">
+          {error && <div className="error-message">{error}</div>}
+
+          <div className="form-group">
+            <label htmlFor="title" className="form-label">
+              Recipe Title
+            </label>
+            <input
+              id="title"
+              type="text"
+              value={formData.title}
+              onChange={(e) =>
+                handleInputChange("title", toTitleCase(e.target.value))
+              }
+              className={`form-input ${validationErrors.title ? "error" : ""}`}
+            />
+            {validationErrors.title && (
+              <span className="field-error">{validationErrors.title}</span>
+            )}
+          </div>
+
+          <div className="form-group">
+            <label htmlFor="category" className="form-label">
+              Category
+            </label>
+            <select
+              id="category"
+              value={formData.category}
+              onChange={(e) => handleInputChange("category", e.target.value)}
+              className={`form-input ${
+                validationErrors.category ? "error" : ""
+              }`}
+            >
+              <option value="" disabled>
+                Select a category
+              </option>
+              {categories
+                ?.filter((cat) => cat.toLowerCase() !== "alle rezepte")
+                .map((cat) => (
+                  <option key={cat} value={cat}>
+                    {cat}
+                  </option>
+                ))}
+            </select>
+            {validationErrors.category && (
+              <span className="field-error">{validationErrors.category}</span>
+            )}
+          </div>
+
+          <div className="form-group">
+            <label htmlFor="servings" className="form-label">
+              Servings
+            </label>
+            <input
+              id="servings"
+              type="number"
+              min="1"
+              value={formData.servings}
+              onChange={(e) => handleInputChange("servings", e.target.value)}
+              className="form-input"
+            />
+          </div>
+
+          <div className="form-group">
+            <label htmlFor="image_url" className="form-label">
+              Image URL
+            </label>
+            <input
+              id="image_url"
+              type="url"
+              value={formData.image_url}
+              onChange={(e) => handleInputChange("image_url", e.target.value)}
+              className="form-input"
+            />
+          </div>
+
+          <div className="form-group">
+            <div className="ingredients-header">
+              <label className="form-label">Ingredients</label>
+              <button type="button" onClick={addIngredient} className="add-btn">
+                + Add Ingredient
+              </button>
+            </div>
+
+            {validationErrors.ingredients && (
+              <span className="field-error">
+                {validationErrors.ingredients}
+              </span>
+            )}
+
+            <div className="ingredients-list">
+              {formData.ingredients.map((ingredient) => (
+                <div key={ingredient.tempId} className="ingredient-row">
+                  <input
+                    id={`ingredient-name-${ingredient.tempId}`}
+                    type="text"
+                    value={ingredient.name || ""}
+                    onChange={(e) =>
+                      handleIngredientChange(
+                        ingredient.tempId,
+                        "name",
+                        e.target.value.toLowerCase()
+                      )
+                    }
+                    className="ingredient-name"
+                    placeholder="Ingredient name"
+                  />
+                  <input
+                    id={`ingredient-quantity-${ingredient.tempId}`}
+                    type="number"
+                    min="0"
+                    step="0.01"
+                    value={ingredient.quantity}
+                    onChange={(e) =>
+                      handleIngredientChange(
+                        ingredient.tempId,
+                        "quantity",
+                        e.target.value
+                      )
+                    }
+                    className="ingredient-quantity"
+                    placeholder="Qty"
+                  />
+                  <select
+                    id={`ingredient-unit-${ingredient.tempId}`}
+                    value={ingredient.unit}
+                    onChange={(e) =>
+                      handleIngredientChange(
+                        ingredient.tempId,
+                        "unit",
+                        e.target.value
+                      )
+                    }
+                    className={`form-input ${
+                      validationErrors.unit ? "error" : ""
+                    }`}
+                  >
+                    {unitOptions.map((option) => (
+                      <option
+                        key={option.value}
+                        value={option.value}
+                        disabled={option.disabled}
+                      >
+                        {option.label}
+                      </option>
+                    ))}
+                  </select>
+
+                  <input
+                    id={`ingredient-notes-${ingredient.tempId}`}
+                    type="text"
+                    value={ingredient.notes}
+                    onChange={(e) =>
+                      handleIngredientChange(
+                        ingredient.tempId,
+                        "notes",
+                        e.target.value
+                      )
+                    }
+                    className="ingredient-notes"
+                    placeholder="Notes (optional)"
+                  />
+                  {formData.ingredients.length > 1 && (
+                    <button
+                      type="button"
+                      onClick={() => removeIngredient(ingredient.tempId)}
+                      className="remove-btn"
+                    >
+                      <X />
+                    </button>
+                  )}
+                </div>
+              ))}
+            </div>
+          </div>
+
+          <div className="form-group">
+            <div className="instructions-header">
+              <label className="form-label">Instructions</label>
+              <button
+                type="button"
+                onClick={addInstruction}
+                className="add-btn"
+              >
+                + Add Step
+              </button>
+            </div>
+
+            {validationErrors.instructions && (
+              <span className="field-error">
+                {validationErrors.instructions}
+              </span>
+            )}
+
+            <div className="instructions-list">
+              {formData.instructions.map((instruction, index) => (
+                <div key={index} className="instruction-row">
+                  <span className="step-number">{index + 1}.</span>
+                  <textarea
+                    value={instruction}
+                    onChange={(e) =>
+                      handleInstructionChange(index, e.target.value)
+                    }
+                    className="instruction-textarea"
+                    rows="2"
+                    onKeyDown={handleEnter}
+                  />
+                  {formData.instructions.length > 1 && (
+                    <button
+                      type="button"
+                      onClick={() => removeInstruction(index)}
+                      className="remove-btn"
+                    >
+                      <X />
+                    </button>
+                  )}
+                </div>
+              ))}
+            </div>
+          </div>
+
+          <div className="form-actions">
+            <button
+              type="button"
+              onClick={handleCancel}
+              className="secondary-btn"
+            >
+              Cancel
+            </button>
+            <button type="submit" disabled={loading} className="primary-btn">
+              {loading
+                ? isEditMode
+                  ? "Updating..."
+                  : "Creating..."
+                : isEditMode
+                ? "Update Recipe"
+                : "Create Recipe"}
+            </button>
+          </div>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default RecipeForm;
